Better int types, Tests, Travis CI

+++ src/common/rolling_index_test.go
@@ -8,10 +8,10 @@ import (
 
 func TestRollingIndex(t *testing.T) {
 	size := 10
-	testSize := 3 * size
+	testSize := int64(3 * size)
 	RollingIndex := NewRollingIndex("test", size)
 	items := []string{}
-	for i := 0; i < testSize; i++ {
+	for i := int64(0); i < testSize; i++ {
 		item := fmt.Sprintf("item%d", i)
 		RollingIndex.Set(item, i)
 		items = append(items, item)
@@ -23,10 +23,10 @@ func TestRollingIndex(t *testing.T) {
 		t.Fatalf("lastIndex should be %d, not %d", expectedLastIndex, lastIndex)
 	}
 
-	start := (testSize / (2 * size)) * (size)
+	start := (testSize / (2 * int64(size))) * int64(size)
 	count := testSize - start
 
-	for i := 0; i < count; i++ {
+	for i := int64(0); i < count; i++ {
 		if cached[i] != items[start+i] {
 			t.Fatalf("cached[%d] should be %s, not %s", i, items[start+i], cached[i])
 		}
@@ -44,7 +44,7 @@ func TestRollingIndex(t *testing.T) {
 
 	var item interface{}
 
-	indexes := []int{10, 17, 29}
+	indexes := []int64{10, 17, 29}
 	for _, i := range indexes {
 		item, err = RollingIndex.GetItem(i)
 		if err != nil {
@@ -61,7 +61,7 @@ func TestRollingIndex(t *testing.T) {
 	}
 
 	//Test updating an item in place
-	updateIndex := 26
+	updateIndex := int64(26)
 	updateValue := "Updated Item"
 
 	err = RollingIndex.Set(updateValue, updateIndex)
@@ -80,7 +80,7 @@ func TestRollingIndex(t *testing.T) {
 
 func TestRollingIndexSkip(t *testing.T) {
 	size := 10
-	testSize := 25
+	testSize := int64(25)
 	RollingIndex := NewRollingIndex("test", size)
 
 	_, err := RollingIndex.Get(-1)
@@ -89,7 +89,7 @@ func TestRollingIndexSkip(t *testing.T) {
 	}
 
 	items := []string{}
-	for i := 0; i < testSize; i++ {
+	for i := int64(0); i < testSize; i++ {
 		item := fmt.Sprintf("item%d", i)
 		RollingIndex.Set(item, i)
 		items = append(items, item)
@@ -99,7 +99,7 @@ func TestRollingIndexSkip(t *testing.T) {
 		t.Fatalf("Skipping index 0 should return ErrTooLate")
 	}
 
-	skipIndex1 := 9
+	skipIndex1 := int64(9)
 	expected1 := items[skipIndex1+1:]
 	cached1, err := RollingIndex.Get(skipIndex1)
 	if err != nil {
@@ -113,7 +113,7 @@ func TestRollingIndexSkip(t *testing.T) {
 		t.Fatalf("expected and cached not equal")
 	}
 
-	skipIndex2 := 15
+	skipIndex2 := int64(15)
 	expected2 := items[skipIndex2+1:]
 	cached2, err := RollingIndex.Get(skipIndex2)
 	if err != nil {
@@ -127,7 +127,7 @@ func TestRollingIndexSkip(t *testing.T) {
 		t.Fatalf("expected and cached not equal")
 	}
 
-	skipIndex3 := 27
+	skipIndex3 := int64(27)
 	expected3 := []string{}
 	cached3, err := RollingIndex.Get(skipIndex3)
 	if err != nil {
