+++ src/node/config.go
@@ -13,7 +13,7 @@ type Config struct {
 	HeartbeatTimeout time.Duration `mapstructure:"heartbeat"`
 	TCPTimeout       time.Duration `mapstructure:"timeout"`
 	CacheSize        int           `mapstructure:"cache-size"`
-	SyncLimit        int           `mapstructure:"sync-limit"`
+	SyncLimit        int64         `mapstructure:"sync-limit"`
 	Logger           *logrus.Logger
 	TestDelay uint64 `mapstructure:"test_delay"`
 }
@@ -21,7 +21,7 @@ type Config struct {
 func NewConfig(heartbeat time.Duration,
 	timeout time.Duration,
 	cacheSize int,
-	syncLimit int,
+	syncLimit int64,
 	logger *logrus.Logger) *Config {
 
 	return &Config{
